Update utils from onemen (TabMixPlus)

+++ chrome/utils/xPref.sys.mjs
@@ -5,17 +5,14 @@ export const xPref = {
   // testei com tipos complexos como nsIFile, não sei como detectar
   // uma preferência assim, na verdade nunca vi uma
   get: function (prefPath, def = false, valueIfUndefined, setDefault = true) {
-    let sPrefs = def ?
-                   Services.prefs.getDefaultBranch(null) :
-                   Services.prefs;
+    const sPrefs = def ? Services.prefs.getDefaultBranch(null) : Services.prefs;
 
     try {
       switch (sPrefs.getPrefType(prefPath)) {
         case 0:
           if (valueIfUndefined != undefined)
             return this.set(prefPath, valueIfUndefined, setDefault);
-          else
-            return undefined;
+          else return undefined;
         case 32:
           return sPrefs.getStringPref(prefPath);
         case 64:
@@ -23,16 +20,14 @@ export const xPref = {
         case 128:
           return sPrefs.getBoolPref(prefPath);
       }
-    } catch (ex) {
+    } catch {
       return undefined;
     }
     return;
   },
 
   set: function (prefPath, value, def = false) {
-    let sPrefs = def ?
-                   Services.prefs.getDefaultBranch(null) :
-                   Services.prefs;
+    const sPrefs = def ? Services.prefs.getDefaultBranch(null) : Services.prefs;
 
     switch (typeof value) {
       case 'string':
@@ -46,10 +41,9 @@ export const xPref = {
   },
 
   lock: function (prefPath, value) {
-    let sPrefs = Services.prefs;
+    const sPrefs = Services.prefs;
     this.lockedBackupDef[prefPath] = this.get(prefPath, true);
-    if (sPrefs.prefIsLocked(prefPath))
-      sPrefs.unlockPref(prefPath);
+    if (sPrefs.prefIsLocked(prefPath)) sPrefs.unlockPref(prefPath);
 
     this.set(prefPath, value, true);
     sPrefs.lockPref(prefPath);
@@ -59,11 +53,9 @@ export const xPref = {
 
   unlock: function (prefPath) {
     Services.prefs.unlockPref(prefPath);
-    let bkp = this.lockedBackupDef[prefPath];
-    if (bkp == undefined)
-      Services.prefs.deleteBranch(prefPath);
-    else
-      this.set(prefPath, bkp, true);
+    const bkp = this.lockedBackupDef[prefPath];
+    if (bkp == undefined) Services.prefs.deleteBranch(prefPath);
+    else this.set(prefPath, bkp, true);
   },
 
   clear: Services.prefs.clearUserPref,
@@ -75,12 +67,12 @@ export const xPref = {
   addListener: function (prefPath, trat) {
     this.observer = function (aSubject, aTopic, prefPath) {
       return trat(xPref.get(prefPath), prefPath);
-    }
+    };
 
     Services.prefs.addObserver(prefPath, this.observer);
     return {
       prefPath,
-      observer: this.observer
+      observer: this.observer,
     };
   },
 
@@ -88,5 +80,5 @@ export const xPref = {
   // Só precisa passar a var definida quando adicionou
   removeListener: function (obs) {
     Services.prefs.removeObserver(obs.prefPath, obs.observer);
-  }
-}
+  },
+};
